Securely deleting files
Added an option to securely delete files. This is achieved by using the
shred command (installed by default on ubuntu servers).

Tested by manually making a hard link of a file, when the option is
enabled the linked file is destroyed as well. When secure deletion is
disabled the linked file still has the original contents.